requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
grpcio==1.27.2
grpcio-tools==1.27.2
idna==2.6
jsonpath==0.82
more-itertools==8.2.0
pika==1.1.0
pluggy==0.6.0
protobuf==3.7.0rc3
py==1.5.3
pymongo==3.6.1
pymssql==2.1.4
pytest==3.6.0
pytest-html==1.19.0
pytest-metadata==1.8.0
redis==3.4.1
requests==2.18.4
six==1.12.0
urllib3==1.22
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
